requirements.txt
#requriements

# Core dependencies
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
email-reply-parser>=0.5.12
pytz>=2024.1

# AI and processing
langchain-openai>=0.0.5
langchain>=0.3.20
langchain-core>=0.3.41
python-dotenv>=1.0.0

# Development dependencies
pytest>=8.0.0
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0
